+++ shilling.sol
@@ -4,7 +4,7 @@ pragma solidity = 0.8.4;
 interface TSHDATA {
     function totalSupply() external view returns (uint80);
     function balanceOf(address account, address m_sender) external view returns (uint256);
-    function allowance(address owner, address spender) external view returns (uint80);
+    function allowance(address owner, address spender) external view returns (uint256);
     function sendLiquid(address from, address to, uint256 amount, address m_sender) external returns (bool);
     function approve(address spender, uint256 amount, address m_sender, uint8 _type) external returns (bool);
     function mint(address to, uint256 value) external returns (bool);
